+++ tests/MockHelper.php
@@ -4,21 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Security;
 
-/**
- * Mock for the hash_hmac() function
- *
- * @param $algo
- * @param $data
- * @param $key
- * @param bool $raw_output
- *
- * @return string
- */
-function hash_hmac($algo, $data, $key, $raw_output = false)
-{
-    return MockHelper::$mock_hash_hmac ?? hash_hmac($algo, $data, $key, $raw_output);
-}
-
 function extension_loaded($name)
 {
     return MockHelper::$mock_extension_loaded ?? \extension_loaded($name);
@@ -37,11 +22,6 @@ function openssl_decrypt($data, $method, $password, $options = 1, $iv = '')
 class MockHelper
 {
     /**
-     * @var bool|string|null value to be returned by mocked hash_hmac() function.
-     * null means normal hash_hmac() behavior.
-     */
-    public static $mock_hash_hmac;
-    /**
      * @var bool|null value to be returned by mocked extension_loaded() function.
      * null means normal extension_loaded() behavior.
      */
@@ -59,7 +39,6 @@ class MockHelper
 
     public static function resetMocks(): void
     {
-        static::$mock_hash_hmac = null;
         static::$mock_extension_loaded = null;
         static::$mock_openssl_encrypt = null;
         static::$mock_openssl_decrypt = null;